// SDKs/twilio.com/twilio_serverless_v1/1.37.3/typescript/src/sdk/models/shared/serverlessv1serviceenvironmentdeployment.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class ServerlessV1ServiceEnvironmentDeployment extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=build_sid" })
  buildSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=environment_sid" })
  environmentSid?: string;

  @SpeakeasyMetadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}